introduce StandardMaterialBuilder
