+++ src/Test/index.ts
@@ -6,7 +6,7 @@ import "@babylonjs/core/Engines/WebGPU/Extensions/engine.renderTarget";
 import "@babylonjs/core/Engines/WebGPU/Extensions/engine.renderTargetTexture";
 
 import { BaseRuntime } from "./baseRuntime";
-import { SceneBuilder } from "./Scene/wasmPhysicsTestScene";
+import { SceneBuilder } from "./Scene/shadingTestScene";
 
 const canvas = document.createElement("canvas");
 canvas.style.width = "100%";
